Fix yet another bug in Make 3.81 ...
Make 3.81 on Debian can't tell the difference between the phony target
"build" and the directory target "build/" and so complains about a spurious
"circular dependency". This commit renames the phony target to "generate"
as a workaround, until Debian wake up and realise that "stable" and
"avoiding any and all patching" are not the same thing.